+++ src/koa-middleware/auth/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  authenticate: require('./authenticate'),
+  parseAuthToken: require('./parseAuthToken'),
+  optionalAuthenticate: require('./optionalAuthenticate'),
+};
